redux selected id, split styles to components, tab buttons
