+++ packages/client/src/redux/store.tsx
@@ -15,6 +15,7 @@ import draggedPropRowSlice from "./features/rowDnd/draggedPropRowSlice";
 import draggedActantRowSlice from "./features/rowDnd/draggedActantRowSlice";
 import rowsExpandedSlice from "./features/statementList/rowsExpandedSlice";
 import draggedRowIdSlice from "./features/statementList/draggedRowIdSlice";
+import selectedDetailIdSlice from "./features/entityDetail/selectedDetailIdSlice";
 
 const store: Store = configureStore({
   reducer: {
@@ -27,6 +28,9 @@ const store: Store = configureStore({
     statementList: combineReducers({
       rowsExpanded: rowsExpandedSlice,
       draggedRowId: draggedRowIdSlice,
+    }),
+    entityDetail: combineReducers({
+      selectedDetailId: selectedDetailIdSlice,
     }),
     rowDnd: combineReducers({
       draggedPropRow: draggedPropRowSlice,
